Inline Game class into main module
